Move bluetooth related functions to minimalin.c
